fix error TodoList
