--- src/App/components/ToDoList/TodoList.mock.ts
@@ -1,24 +1,0 @@
-import faker from 'faker';
-
-export function getFakeData() {
-	return {
-		items: [
-			{
-				id: faker.random.uuid(),
-				text: faker.random.words()
-			},
-			{
-				id: faker.random.uuid(),
-				text: faker.random.words()
-			},
-			{
-				id: faker.random.uuid(),
-				text: faker.random.words()
-			},
-			{
-				id: faker.random.uuid(),
-				text: faker.random.words()
-			}
-		]
-	};
-}
