--- src/App/components/ToDoList/TodoList.st.css
@@ -1,4 +1,0 @@
-.root {
-	display: flex;
-	justify-content: center;
-}
