+++ src/App/components/TodoList/TodoList.mock.ts
@@ -1,0 +1,19 @@
+import faker from 'faker';
+
+export function getFakeData() {
+	return {
+		items: [{
+			id: faker.random.uuid(),
+			text: faker.random.words()
+		},      {
+			id: faker.random.uuid(),
+			text: faker.random.words()
+		},      {
+			id: faker.random.uuid(),
+			text: faker.random.words()
+		},    {
+			id: faker.random.uuid(),
+			text: faker.random.words()
+		}]
+	};
+}
