+++ src/App/components/TodoList/TodoList.st.css
@@ -1,0 +1,3 @@
+.root {
+	margin: 0 auto;
+}
